In Cucumber, if there is an exception in the beforeScenario handler, the afterScenario is not called. Catch the exception in the beforeScenario and free the device in such a case.
Longer version: For each new test that starts running, ATD marks a device "busy" before creating a new driver instance from the beforeScenario. When running a suite of tests, if there is a problem in the driver initialisation resulting in an exception for each test, eventually all devices end up getting marked as "BUSY" in beforeScenario. Once all the devices are "busy", for subsequent tests, there are no available devices, and the thread ends up in suspended state till some device becomes "free". The solution is to catch the exception when creating the driver, and in this case, free the device before re-throwing the exception.